# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "ClipResult": {
        "description": "Result of creating a clip from a source.",
        "properties": {
          "clip_id": {
            "title": "Clip Id",
            "type": "string"
          },
          "duration_s": {
            "title": "Duration S",
            "type": "number"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "fps": {
            "title": "Fps",
            "type": "number"
          },
          "has_audio": {
            "title": "Has Audio",
            "type": "boolean"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "thumbnail_path": {
            "title": "Thumbnail Path",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "clip_id",
          "filename",
          "duration_s",
          "width",
          "height",
          "fps",
          "has_audio",
          "thumbnail_path"
        ],
        "title": "ClipResult",
        "type": "object"
      },
...
      "CreateClipRequest": {
        "properties": {
          "dataset_dir": {
            "title": "Dataset Dir",
            "type": "string"
          },
          "end_s": {
            "title": "End S",
            "type": "number"
          },
          "keep_audio": {
            "default": true,
            "title": "Keep Audio",
            "type": "boolean"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "start_s": {
            "title": "Start S",
            "type": "number"
          },
          "target_fps": {
            "default": 24,
            "title": "Target Fps",
            "type": "integer"
          },
          "target_longest_side": {
            "default": 1280,
            "title": "Target Longest Side",
            "type": "integer"
          }
        },
        "required": [
          "source_path",
          "dataset_dir",
          "start_s",
          "end_s"
        ],
        "title": "CreateClipRequest",
        "type": "object"
      },
      "CreateClipsBatchRequest": {
        "properties": {
          "dataset_dir": {
            "title": "Dataset Dir",
            "type": "string"
          },
          "keep_audio": {
            "default": true,
            "title": "Keep Audio",
            "type": "boolean"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/ClipSegment"
            },
            "title": "Segments",
            "type": "array"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "target_fps": {
            "default": 24,
            "title": "Target Fps",
            "type": "integer"
          },
          "target_longest_side": {
            "default": 1280,
            "title": "Target Longest Side",
            "type": "integer"
          }
        },
        "required": [
          "source_path",
          "dataset_dir",
          "segments"
        ],
        "title": "CreateClipsBatchRequest",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
...
      "UpdateCaptionRequest": {
        "properties": {
          "caption": {
            "title": "Caption",
            "type": "string"
          },
          "clip_id": {
            "title": "Clip Id",
            "type": "string"
          },
          "dataset_dir": {
            "title": "Dataset Dir",
            "type": "string"
          }
        },
        "required": [
          "dataset_dir",
          "clip_id",
          "caption"
        ],
        "title": "UpdateCaptionRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenLTX Trainer Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dataset/clips": {
      "post": {
        "operationId": "route_create_clip_api_dataset_clips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Create Clip",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/clips/batch": {
      "post": {
        "operationId": "route_create_clips_batch_api_dataset_clips_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClipsBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClipResult"
                  },
                  "title": "Response Route Create Clips Batch Api Dataset Clips Batch Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Create Clips Batch",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/clips/caption": {
      "post": {
        "operationId": "route_update_caption_api_dataset_clips_caption_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Update Caption",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/clips/delete": {
      "post": {
        "operationId": "route_delete_clip_api_dataset_clips_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteClipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Delete Clip",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/clips/delete-all": {
      "post": {
        "operationId": "route_delete_all_clips_api_dataset_clips_delete_all_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAllClipsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAllClipsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Delete All Clips",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/clips/list": {
      "post": {
        "operationId": "route_list_clips_api_dataset_clips_list_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListClipsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClipRecord"
                  },
                  "title": "Response Route List Clips Api Dataset Clips List Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Clips",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/clips/thumbnail": {
      "post": {
        "operationId": "route_get_thumbnail_api_dataset_clips_thumbnail_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetThumbnailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThumbnailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Get Thumbnail",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/images": {
      "post": {
        "operationId": "route_import_image_api_dataset_images_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Import Image",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/probe": {
      "post": {
        "operationId": "route_probe_source_api_dataset_probe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProbeSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMediaInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Probe Source",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/scan": {
      "post": {
        "operationId": "route_scan_directory_api_dataset_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanDirectoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDirectoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Scan Directory",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/scenes/detect": {
      "post": {
        "operationId": "route_detect_scenes_api_dataset_scenes_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectScenesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SceneProposal"
                  },
                  "title": "Response Route Detect Scenes Api Dataset Scenes Detect Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Detect Scenes",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/stream-video": {
      "get": {
        "description": "Stream a source video file for in-app preview.",
        "operationId": "route_stream_video_api_dataset_stream_video_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Stream Video",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/trigger/audit": {
      "post": {
        "operationId": "route_audit_trigger_api_dataset_trigger_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClipRecord"
                  },
                  "title": "Response Route Audit Trigger Api Dataset Trigger Audit Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Audit Trigger",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/trigger/prepend": {
      "post": {
        "operationId": "route_prepend_trigger_api_dataset_trigger_prepend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrependTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrependTriggerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Prepend Trigger",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/trigger/validate": {
      "post": {
        "operationId": "route_validate_trigger_api_dataset_trigger_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Validate Trigger",
        "tags": [
          "dataset"
        ]
      }
    },
    "/api/dataset/validate": {
      "post": {
        "operationId": "route_validate_dataset_api_dataset_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Validate Dataset",
        "tags": [
          "dataset"
        ]
      }
    },
...
  }
}